requirements.txt
anyio==3.6.2
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cryptography==40.0.2
cssselect==1.2.0
cymem==2.0.7
dnspython==2.3.0
docx2txt==0.8
email-validator==2.0.0.post1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
fastapi==0.95.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.11.0
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
locationtagger==0.0.1
lxml==4.9.2
MarkupSafe==2.1.2
murmurhash==1.0.9
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
orjson==3.8.10
packaging==23.1
pandas==2.0.0
pathy==0.10.1
pdfminer.six==20221105
pdfplumber==0.9.0
phonenumbers==8.13.9
Pillow==9.5.0
platformdirs==3.2.0
preshed==3.0.8
pycodestyle==2.10.0
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
scipy==1.10.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
starlette==0.26.1
stemming==1.0.1
thinc==8.1.9
tika==2.6.0
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
uritools==4.0.1
urlextract==1.8.0
urllib3==1.26.15
uvicorn==0.21.1
Wand==0.6.11
wasabi==1.1.1
watchfiles==0.19.0
websockets==11.0.1
